# ElectronicDiary.WebApi.Client/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ElectronicDiary.Server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Query/studentsInClass": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Query 2: Retrieve information about all students in a specified class by its identifier, sorted by full name",
        "parameters": [
          {
            "name": "classId",
            "in": "query",
            "description": "Class identifier",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Student"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Student"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Student"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Query/studentsByDate": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Query 3: Retrieve information about all students who received grades on a specified date",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "description": "Date of the grades",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentGradesDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentGradesDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentGradesDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Query/studentsWithMaxAverageByPeriod": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Query 5: Retrieve students with the highest average grade over a specified period",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "description": "Start date of the period",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "End date of the period",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentAverageGradeDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentAverageGradeDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentAverageGradeDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Class": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "letters": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GradeType": {
        "enum": [
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
      "Student": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          },
          "patronymic": {
            "type": "string",
            "nullable": true
          },
          "class": {
            "$ref": "#/components/schemas/Class"
          },
          "birthday": {
            "type": "string",
            "format": "date"
          },
          "passport": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StudentAverageGradeDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "integer",
            "description": "Student's identifier",
            "format": "int32"
          },
          "surname": {
            "type": "string",
            "description": "Student's surname",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "Student's first name",
            "nullable": true
          },
          "patronymic": {
            "type": "string",
            "description": "Student's patronymic",
            "nullable": true
          },
          "averageGrade": {
            "type": "number",
            "description": "Student's average grade",
            "format": "double"
          }
        },
        "additionalProperties": false,
        "description": "DTO for representing a student's average grade"
      },
...
      "StudentGradesDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "integer",
            "description": "Student's identifier",
            "format": "int32"
          },
          "surname": {
            "type": "string",
            "description": "Student's surname",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "Student's first name",
            "nullable": true
          },
          "patronymic": {
            "type": "string",
            "description": "Student's patronymic",
            "nullable": true
          },
          "subject": {
            "type": "string",
            "description": "Name of the subject for which the grade was received",
            "nullable": true
          },
          "grade": {
            "$ref": "#/components/schemas/GradeType"
          },
          "date": {
            "type": "string",
            "description": "Date when the grade was received",
            "format": "date"
          }
        },
        "additionalProperties": false,
        "description": "DTO for representing information about a student's grades on a specific day"
      },
...
    }
  }
}